edit folder functions and convention
