+++ exercise-reactjs/src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
+import '../src/stylesheet/style.css';
 import ArticleList from './app/pages/article/containers/ArticleList';
 import { BrowserRouter as Router } from "react-router-dom";
 import { Route, Switch } from "react-router";
-import '../src/stylesheet/style.css';
 import ArticleDetail from './app/pages/article/containers/ArticleDetail';
 
 function App() {
