+++ exercise-reactjs/src/app/core/functions/text.ts
@@ -1,0 +1,3 @@
+export const replaceText = (text: string, value: string) => {
+  return text.replace('{value}', value);
+}
